+++ src/ts/static/linksIcons.ts
@@ -1,0 +1,28 @@
+const linksIcons = [
+    {
+        linkTitle: 'The PHP Foundation',
+        iconName: 'phpfoundation.png',
+    },
+    {
+        linkTitle: 'Upcoming conferences',
+        iconName: 'talk.png',
+    },
+    {
+        linkTitle: 'User Group Events',
+        iconName: 'events.png',
+    },
+    {
+        linkTitle: 'Special Thanks',
+        iconName: 'thanks.png',
+    },
+    {
+        linkTitle: '@official_php',
+        iconName: 'x.svg',
+    },
+    {
+        linkTitle: '@php@fosstodon.org',
+        iconName: 'mastodon.png',
+    },
+]
+
+export default linksIcons