+++ src/ts/static/socialLinks.ts
@@ -1,0 +1,11 @@
+import type { HomeLink } from '@/types'
+
+const recommendedLinks: HomeLink[] = [
+    {
+        href: 'https://www.linkedin.com/company/phpnet/',
+        iconName: 'linkedin.svg',
+        title: '@phpnet',
+    },
+]
+
+export default recommendedLinks