+++ src/ts/templates/homeLinkTemplate.ts
@@ -1,42 +1,34 @@
 import type { HomeLink } from '@/types'
 import getImageUrl from '@/modules/getImageUrl'
 
-export default (link: HomeLink): Element => {
+export default (link: HomeLink): HTMLElement => {
     const src = getImageUrl(`icons/${link.iconName}`)
-
     const img = createImage(src)
     const span = createSpan(link.title)
-    const a = createLink(img, span, link.link)
 
-    return createParagraph(a)
+    return createAnchor(img, span, link.href)
 }
 
-function createParagraph(a: Element): Element {
-    const p = document.createElement('p')
-    p.classList.add('panel', 'php-revival-panel')
-    p.appendChild(a)
-
-    return p
-}
-
-function createImage(src: string): Element {
+function createImage(src: string): HTMLImageElement {
     const img = document.createElement('img')
     img.src = src
 
     return img
 }
 
-function createSpan(title: string): Element {
+function createSpan(title: string): HTMLSpanElement {
     const span = document.createElement('span')
     span.textContent = title
 
     return span
 }
 
-function createLink(img: Element, span: Element, href: string): Element {
+function createAnchor(img: Element, span: Element, href: string): HTMLAnchorElement {
     const a = document.createElement('a')
+
     a.href = href
     a.target = '_blank'
+    a.className = 'php-revival-home-links-section__link'
 
     a.appendChild(img)
     a.appendChild(span)